Refactor: map: hooks cleanup + neighbs auto-zoom into its component
